condicionais_repetição_vetores_10/05

+++ condicionais_repetição_vetores_10_05/exercicio3_produto/Program.cs
@@ -13,7 +13,7 @@ namespace exercicio3_produto
             double PreçoUnitário = double.Parse(Console.ReadLine());
 
             Console.WriteLine("\nQuantas unidades do produto você adquiriu?");
-            double Quantidade = double.Parse(Console.ReadLine());
+            int Quantidade = int.Parse(Console.ReadLine());
 
             double Total = PreçoUnitário * Quantidade;
